+++ ledger/ledger-on-memory/BUILD.bazel
@@ -5,7 +5,7 @@ load(
     "//bazel_tools:scala.bzl",
     "da_scala_binary",
     "da_scala_library",
-    "da_scala_test",
+    "da_scala_test_suite",
 )
 load("//ledger/ledger-api-test-tool:conformance.bzl", "conformance_test")
 load("@io_bazel_rules_docker//java:image.bzl", "java_image")
@@ -38,7 +38,33 @@ da_scala_library(
     ],
 )
 
-da_scala_test(
+da_scala_library(
+    name = "ledger-on-memory-test-lib",
+    srcs = glob(["src/test/lib/scala/**/*.scala"]),
+    deps = [
+        ":ledger-on-memory",
+        "//daml-lf/data",
+        "//daml-lf/engine",
+        "//language-support/scala/bindings",
+        "//ledger-api/rs-grpc-bridge",
+        "//ledger-api/testing-utils",
+        "//ledger/caching",
+        "//ledger/ledger-api-common",
+        "//ledger/ledger-api-health",
+        "//ledger/metrics",
+        "//ledger/participant-state",
+        "//ledger/participant-state/kvutils",
+        "//ledger/participant-state/kvutils:kvutils-tests-lib",
+        "//libs-scala/contextualized-logging",
+        "//libs-scala/resources",
+        "@maven//:com_typesafe_akka_akka_actor_2_12",
+        "@maven//:com_typesafe_akka_akka_stream_2_12",
+        "@maven//:org_scalactic_scalactic_2_12",
+        "@maven//:org_scalatest_scalatest_2_12",
+    ],
+)
+
+da_scala_test_suite(
     name = "ledger-on-memory-tests",
     size = "small",
     srcs = glob(["src/test/suite/**/*.scala"]),
@@ -48,12 +74,11 @@ da_scala_test(
     resources = glob(["src/test/resources/*"]),
     deps = [
         ":ledger-on-memory",
+        ":ledger-on-memory-test-lib",
         "//daml-lf/data",
-        "//daml-lf/engine",
         "//language-support/scala/bindings",
         "//ledger-api/rs-grpc-bridge",
         "//ledger-api/testing-utils",
-        "//ledger/caching",
         "//ledger/ledger-api-common",
         "//ledger/ledger-api-health",
         "//ledger/metrics",
